pywatchman: check for pid being None before killing it
On Python 3 there seem to be weird edge cases where the pid can be None even
if we hold a reference to proc.

+++ python/pywatchman/__init__.py
@@ -437,7 +437,8 @@ class CLIProcessTransport(Transport):
 
     def close(self):
         if self.proc:
-            self.proc.kill()
+            if self.proc.pid is not None:
+                self.proc.kill()
             self.proc = None
 
     def _connect(self):
